src/main/webapp/app/entities/nonce/nonce.service.ts
import axios from 'axios';

import { INonce } from '@/shared/model/nonce.model';

const baseApiUrl = 'api/nonces';

export default class NonceService {
  public find(id: number): Promise<INonce> {
    return new Promise<INonce>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id: number): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity: INonce): Promise<INonce> {
    return new Promise<INonce>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity: INonce): Promise<INonce> {
    return new Promise<INonce>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
